adds dependencies
now the makefile knows that all .c files depend on the hellomake.h file. so now, when the hellomake.h file is changed, make will recompile the .c files.

we create the DEPS macro, which is a set of .h files on which the .c files depend. then we create a rule, that applies to all files ending in the .o suffix. the rule says that the .o file depends upon the .c version of the file and the .h files included in the DEPS macro. the rule then says that to generate the .o file, make needs to compile the .c file using the compiler defined in the CC macro. the -c flag says to generate the object file, the -o $@ says to put the output of the compilation in the file named on the left side of the :, the $< is the first item in the dependencies list, and the CFLAGS macro is defined above.