adds new directiories
this makefile adds paths to the include and lib directories, and places the object files in an obj subdirectory within the src directory. it also has a macro defined for any libraries that must be included, such as the math library -lm. this makefile should be located in the src directory. it also includes a rule for cleaning up the source and object directories when 'make clean' is typed. the .PHONY rule keeps make from doing something with a file named 'clean'